Further Changes Added
New modules added Plusones, RSVPs, Songs, Messages and update Guests.
Guests contains the details for them to login.
The Plusones will link back to the Guest table as the Guests will be the
lead guests and the plusones will be the plusones! CRUD all done, need
to lay down some further logic for the site to work and also make it
look pretty.