Logged in users are updating correct modules
Managed to correctly set the correct guest_id based upon the login
account that they are mapped to. However it is being shown as a select
box and I need to make it hidden.
Also fixed the layout problem where the edges of the side bar are now
layed out correctly.
More static data has been added.