More images added
Adding more images, using images for the menu for now to make the text
seem more generic against all browsers.
Added the ability for Admins to see certain actions and for Authors to
only do certain actions.
Hid the users selection box as I don't want users to be able to see who
else is attending the wedding or have the ability to update on their
behalf
Removing code that is not required
General cosmetic changes