Joining models together
Able to join two models on a single view so that we can perform the add.
Tweaking the RSVP add functionality so that we can pass in multiple
plusones but this still needs work.
Adding success pages for the user to be redirected to upon successful
submission.
Refactoring some code as it wasn't quite what I intended.