Checkbox validation added
RSVP now has the ability to validate the checkboxes selected, so for
example the user can select whether or not to go to either or both of
the ceremony and the banquet but if they select the not attending option
then it deselects all of the other options as you can't not go and go at
the same time.
Obtained the image for the checked option with a relevant image.