Further amendments
Added a logout button to the menu based whether or not a user session is
present.
Amended some text.
Trying to get the adding plusone javascript done - trying to extra text
inputs when the add function is requested - still working progress.
Refined some of the logic on the check boxes to disable buttons and
re-enable them when they are not going etc...