Tweaks
Trying to implement a check that prevents authors from viewing certain
areas or trying to access certain areas of the site.
Tweaking the JavaScript on the RSVP page so that when adding extra
guests it will stop them from going over 5 extra plus ones (want to
include kids if they have any). Need to loop round the input elements to
count them then check if they are over 5 if so then disable the add
button and throw a message.