RSVP is now passing in multiple records
Now able to pass in more than one plusone into the database. Not the
best way of doing it really but it works.
Also revised the javascript and sorting out bits of the authorisation
problem that I've been working to prevent certain users access to
certain pages.
The javascript will need to be amended when removing previous input text
boxes.