Access Denied
Now managed to lock everything down based on permission, so anyone that
isn't a user cannot see anything except the home page which only allows
the user to see the login page.
When they are logged in they are only allowed to do certain actions (add
or logout).
Admins can do everything.
Also the flash message needs to be coloured differently so that on
successfully actions it should display a green colour and if it is
anything else then red.