Tweaking RSVP Validation
Managed to update the validation for the checkboxes so that it forces
the user to select at least one checkbox so that it prevents crap data
being added to the database.
Added the Modal dialog instead of using the alert boxes as it is just
that nicer. This needs to be styled correctly so that it is consistent.