UI Tweaks
Finally got rid of the orange ring around the focused elements. This was
driving me crazy as I should have tested it in other browsers to see if
the problem was related to Chrome.
Also got the button to correctly show the default and hover images for
the close button on the dialog box.