Rsvp modifications
Changing some of the logic, if the guest has not RSVP'd then when they
click on the link to RSVP they are redirected to the add action. If they
have already RSVP'd then redirect them to the view action to see what
they have done and to also edit their response.