RSVP Summary ...
The summary page for the RSVP module is now done, it will display a
different message depending on what the user has chosen in their reply.
Also when they decide to edit their response it will pre-populate the
fields so that they can update accordingly.
The problem is that the plusones aren't populating so this will need
more attention.