Edit RSVP
Trying to get the fields pre-populated with the data that has already
been saved when the user has made an RSVP so that they can clearly see
what they need to edit. The checkboxes are correct but the plusones need
to be populated. Committing my non working code for now so that I can
carry on tomorrow.