Fixes
Noticed that when a user is within the RSVP area and then decides to
logout and then decides to log back in they get redirected to add
another RSVP record. If they try and add a record they will get a
constraint as only one RSVP record should exist for a guest.
To resolve this I check to see if the RSVP record exists for the guest
if it does then redirect the guest to the view record. Otherwise carry
on as if they need to add.
I have repeated code so will need to put it into a function so that it
can be shared. But for the time being this needs to go.