Twitter Wall Added
Twitter wall added, taken from Remy Sharp's repo
(https://github.com/remy/twitterwall) as we need to display peoples twit
pics on the day of the wedding. Need to update the styles to match our
wedding site. Also adding a reporting page to give us a quick summary of
the guests that will be attending.