Twitter Wall - Updates
API has changed as version 1 is now gone (410 error) so will need to
update to v1.1. Having problems as I need to read through the API to see
what has changed. Also making a few cosmetic changes as we will be
adding a background image instead of having the current logo and side
bar.